requirements.txt
streamlit>=1.11.1,!=1.12.0
plotly
pandas
ec3-python-wrapper